LPS-133960 Fix TableReferenceDefinition class name typos
