+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/change/tracking/spi/reference/DDMDataProviderInstanceTableReferenceDefinition.java
@@ -29,7 +29,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Preston Crary
  */
 @Component(service = TableReferenceDefinition.class)
-public class DDMDataProviderInstanceTableReferenceDefinintion
+public class DDMDataProviderInstanceTableReferenceDefinition
 	implements TableReferenceDefinition<DDMDataProviderInstanceTable> {
 
 	@Override
